requirements.txt
absl-py==0.11.0
astor==0.8.1
click==7.1.2
Flask==1.1.2
gast==0.4.0
grpcio==1.34.0
importlib-metadata==3.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
Markdown==3.3.3
MarkupSafe==1.1.1
numpy==1.14.0
Pillow==8.0.1
protobuf==3.14.0
six==1.15.0
tensorboard==1.9.0
tensorflow==1.9.0
termcolor==1.1.0
typing-extensions==3.7.4.3
Werkzeug==1.0.1
zipp==3.4.0
